requirements.txt
coreapi==2.3.0
django==1.11.1
djangorestframework==3.6.3

